Fix social image alt text for Elementor editor
