+++ packages/js/src/containers/FacebookEditor.js
@@ -1,46 +1,14 @@
 /* External dependencies */
 import { compose } from "@wordpress/compose";
 import { withDispatch, withSelect, dispatch as wpDataDispatch } from "@wordpress/data";
-import { validateFacebookImage } from "@yoast/helpers";
-import { FACEBOOK_IMAGE_SIZES, determineFacebookImageMode } from "@yoast/social-metadata-previews";
 
 /* Internal dependencies */
 import FacebookWrapper from "../components/social/FacebookWrapper";
 import getL10nObject from "../analysis/getL10nObject";
 import withLocation from "../helpers/withLocation";
-import { openMedia } from "../helpers/selectMedia";
+import { openMedia, prepareFacebookPreviewImage } from "../helpers/selectMedia";
 
 const socialMediumName = "Facebook";
-
-/**
- * Callback function for selectMedia. Performs actions with the 'image' Object that it gets as an argument.
- *
- * @param {Object} image Object containing data about the selected image.
- *
- * @param {Function} onSelect Callback function received from openMedia. Gets object image' as an argument.
- *
- * @returns {void}
- */
-const imageCallback = ( image ) => {
-	const { width, height } = image;
-	const imageMode = determineFacebookImageMode( { width, height } );
-
-	const idealWidth = FACEBOOK_IMAGE_SIZES[ imageMode + "Width" ];
-	const idealHeight = FACEBOOK_IMAGE_SIZES[ imageMode + "Height" ];
-
-	const idealImageSize = Object.values( image.sizes ).find( size => {
-		return size.width >= idealWidth && size.height >= idealHeight;
-	} );
-
-	const imageUrl = idealImageSize ? idealImageSize.url : image.url;
-
-	wpDataDispatch( "yoast-seo/editor" ).setFacebookPreviewImage( {
-		url: imageUrl,
-		id: image.id,
-		warnings: validateFacebookImage( image ),
-		alt: image.alt || "",
-	} );
-};
 
 /**
  * Lazy function to open the media instance.
@@ -48,7 +16,7 @@ const imageCallback = ( image ) => {
  * @returns {void}
  */
 const selectMedia = () => {
-	openMedia( imageCallback );
+	openMedia( ( image ) => wpDataDispatch( "yoast-seo/editor" ).setFacebookPreviewImage( prepareFacebookPreviewImage( image ) ) );
 };
 
 /* eslint-disable complexity */
